Add Ansible playbook for AVR setup
Add an Ansible playbook which installs the necessary AVR libraries using
apt and pip, and copies the AVRDUDE configuration to the user's home
directory. Also add corresponding instructions in the README for how to
run the Ansible playbook.

Using an Ansible playbook is useful for two reasons: (1) it is
idempotent, which means that it can be run multiple times without
repeating unnecessary steps (and potentially breaking some
configurations), and (2) it can be easily modified to run on remote
machines, which may be helpful in the future.

The current Ansible playbook can only run locally, and must be run from
the repository root directory. If necessary, this can be modified in the
future.